Create Bank_Personal_Loan
This is a classification project where I built an algorithm that will be using historical data to determine which customer is most likely going to take a loan and which is not for the bank to be able to manage their market budget properly and only market loan offer to those that are most likely going to be interested in it. This dataset is obtained from Kaggle and the project is my project in response to the challenge.